// src/app/+core/interfaces/enterprise.interface.ts
export interface IEnterprise {
  birthDate?: string;
  cnpj?: string;
  phone?: string;
  latitude?: string;
  longitude?: string;
  type?: string;
  description?: string;
  google_places_id?: string;
  place_id?: string;
  image_base64?: string;
  neighbourhood?: string;
  zip_code?: string;
  country?: string;
  street?: string;

  number?: number;
  city?: string;
  state?: string;

  id?: number;
  name?: string;
  website?: string;
  address?: any;
  updated_at?: string;
  created_at?: string;
  user?: number;

}